+++ lib/location/index.ts
@@ -2,4 +2,5 @@ export * from "./Location"
 export * from "./Placemark"
 export * from "./Geocoding"
 export * from "./UserLocation"
-export * from "./RecentLocationStorage"
+export * from "./RecentsStorage"
+export * from "./Search"
